角色完成，boss未开始，有动画卡顿问题

+++ bullet-timer-manager.cpp
@@ -34,7 +34,7 @@ void BulletTimeManager::set_status(Status status)
 	this->status = status;
 
 }
-float BulletTimeManager::on_update(int delta)
+float BulletTimeManager::on_update(float delta)
 {
 	float delta_progress = delta * SPEED_PROGRESS;
 	progress += delta_progress * (status == Status::Entering ? 1 : -1);
